refactor webpack to use smaller jquery_bootstrap bundle for badges
