+++ ap/bible_tracker/templates/bible_tracker/index.html
@@ -2,8 +2,6 @@
 {% load staticfiles %}
 {% load bootstrap3 %}
 {% load common_tags %}
-{% bootstrap_css %}
-{% bootstrap_javascript %}
 {% bootstrap_messages %}
 {% block title %}Bible Tracker{% endblock %}
 {% block references %}
