+++ ap/templates/bundles/jquery_bootstrap.js
@@ -1,0 +1,2 @@
+import 'expose-loader?$!expose-loader?jQuery!jquery'
+import 'bootstrap-sass'; import 'bootstrap-sass/assets/stylesheets/_bootstrap.scss'
