+++ package.json
@@ -71,7 +71,7 @@
     "yup": "0.20.0"
   },
   "scripts": {
-    "webpack": "webpack --config webpack/webpack.config.js --progress --colors",
+    "webpack": "webpack --config webpack/webpack.dev.config.js --progress --colors",
     "build": "webpack --config webpack/webpack.prod.config.js --progress --colors",
     "start": "node webpack/server.js"
   },
