+++ webpack/webpack.dev.config.js
@@ -3,7 +3,7 @@ var webpack = require('webpack')
 var BundleTracker  = require('webpack-bundle-tracker')
 var merge = require('webpack-merge')
 
-var commonConfig = require('./webpack.common')
+var commonConfig = require('./webpack.common.config')
 
 var devConfig = {
   entry: {
@@ -11,11 +11,15 @@ var devConfig = {
       'webpack-dev-server/client?http://localhost:3000',
       'webpack/hot/only-dev-server',
     ],
+    jquery_bootstrap: [
+      'webpack-dev-server/client?http://localhost:3000',
+      'webpack/hot/only-dev-server',
+    ],
     attendance: [
       'react-hot-loader/patch',
       'webpack-dev-server/client?http://localhost:3000',
       'webpack/hot/only-dev-server',
-    ]
+    ],
   },
 
   output: {
